# improve performance of formatting timestamps by 35%