Use an optimized routine for formatting timestamps in the default layout for a 35% speedup.

```
goos: darwin
goarch: arm64
pkg: github.com/charlievieth/go-sqlite3/internal/timefmt cpu: Apple M4 Pro
          │   old.txt   │               new.txt               │
          │   sec/op    │   sec/op     vs base                │
Format-14   98.27n ± 1%   63.80n ± 1%  -35.08% (p=0.000 n=10)

          │  old.txt   │            new.txt             │
          │    B/op    │    B/op     vs base            │
Format-14   48.00 ± 0%   48.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

          │  old.txt   │            new.txt             │
          │ allocs/op  │ allocs/op   vs base            │
Format-14   1.000 ± 0%   1.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```